Extract utility function to use in both wrs_mean_test and wrs_mean_test_looped. Revise documentation.
