+++ src/lib.rs
@@ -839,8 +839,8 @@ where
 mod tests {
 
     use super::*;
-    use crate::utils::generate_step_stream;
     use crate::utils::generate_stream_with_constant_probability;
+    use crate::utils::mean_of_means_of_step_stream;
     use std::convert::TryInto;
     use std::io::Read;
     use std::iter;
@@ -1308,69 +1308,33 @@ mod tests {
 
     // For a stream of the form [(0,1),..,(0,1),(1,1),..,(1,1)] with equal numbers
     // of zero and one values and all weights equal to 1, we expect weighted reservoir
-    // sampling to reduce to reservoir sampling (without weights) and thus to produce
+    // sampling to reduce to reservoir sampling (no weights) and thus to produce
     // a reservoir whose mean estimates the mean of the entire stream, which in this case
     // is 0.5. A reservoir sample is generated 50 times. Each time the mean is calculated
     // and the mean of these means is taken. It is asserted that this mean of means is
     // within 5% of the true mean, 0.5.
     //
     // In wrs_mean_test_looped(), the current test (wrs_mean_test()) was run 3000 times
-    // with one failure. Thus we estimate the failure rate to be approximately 1 in 3000.
-    // If this test fails more than once for you, there is likely a problem.
+    // with one failure resulting. Thus we estimate the failure rate to be approximately
+    // 1 in 3000. If this test fails more than once for you, there is likely a problem.
     #[test]
     fn wrs_mean_test() {
-        let stream_length = 100usize;
-        let capacity = stream_length / 2;
-        let num_runs = 50usize;
-        let estimate = |num_runs| {
-            let mut means: Vec<f64> = Vec::with_capacity(capacity);
-            for _i in 0..num_runs {
-                let stream = generate_step_stream(stream_length, capacity, 0i64, 1i64);
-                let stream = wd_iterable(stream, |_x| 1f64);
-                let stream = weighted_reservoir_iterable(stream, capacity, None);
-                let res = last(stream).unwrap();
-                let res: Vec<i64> = res.iter().map(|x| x.value).collect();
-                let mean = res.iter().sum::<i64>() as f64 / capacity as f64;
-                means.push(mean);
-            }
-            let mean_mean = means.iter().sum::<f64>() / num_runs as f64;
-            mean_mean
-        };
-
-        let mean_mean = estimate(num_runs);
-        assert!((mean_mean - 0.5).abs() < 0.05 * 0.5);
+        let mean_means = mean_of_means_of_step_stream();
+        assert!((mean_means - 0.5).abs() < 0.05 * 0.5);
     }
 
-    // wrs_mean_test() (see above) is run 3000 times. In our experience this has lead to one
-    // failure. Thus we estimate the failure rate of wrs_mean_test() to be approximately 1
-    // in 3000.
+    // This test is used to estimate the failure rate of wrs_mean_test (see above).
+    // wrs_mean_test() is run 3000 times. In our experience this has led to one
+    // failure. Thus we estimate the failure rate of wrs_mean_test() to be approximately
+    // 1 in 3000.
     #[test]
     #[ignore]
     fn wrs_mean_test_looped() {
-        let stream_length = 100usize;
-        let capacity = stream_length / 2;
-        let num_runs = 50usize;
-        let estimate = |num_runs| {
-            let mut means: Vec<f64> = Vec::with_capacity(capacity);
-            for _i in 0..num_runs {
-                let stream = generate_step_stream(stream_length, capacity, 0i64, 1i64);
-                let stream = wd_iterable(stream, |_x| 1f64);
-                let stream = weighted_reservoir_iterable(stream, capacity, None);
-                let res = last(stream).unwrap();
-                let res: Vec<i64> = res.iter().map(|x| x.value).collect();
-                let mean = res.iter().sum::<i64>() as f64 / capacity as f64;
-                means.push(mean);
-            }
-            let mean_mean = means.iter().sum::<f64>() / num_runs as f64;
-            mean_mean
-        };
-
-        // The following loop is used to estimate the failure rate of wrs_mean_test().
         let mut failures = 0usize;
         let number_of_runs = 3_000usize;
         for _j in 0..number_of_runs {
-            let mean_mean = estimate(num_runs);
-            if (mean_mean - 0.5).abs() > 0.05 * 0.5 {
+            let mean_means = mean_of_means_of_step_stream();
+            if (mean_means - 0.5).abs() > 0.05 * 0.5 {
                 failures += 1;
             };
         }
